+++ models/CalendarEntry.php
@@ -2,7 +2,9 @@
 
 namespace humhub\modules\calendar\models;
 
+use humhub\modules\calendar\CalendarUtils;
 use humhub\modules\calendar\permissions\ManageEntry;
+use humhub\modules\calendar\widgets\WallEntry;
 use humhub\modules\content\models\Content;
 use humhub\modules\search\interfaces\Searchable;
 use Yii;
@@ -34,14 +36,10 @@ use humhub\modules\user\models\User;
  */
 class CalendarEntry extends ContentActiveRecord implements Searchable
 {
-
-    // Atm not attach meetings to wall
-    public $autoAddToWall = true;
-
     /**
      * @inheritdoc
      */
-    public $wallEntryClass = "humhub\modules\calendar\widgets\WallEntry";
+    public $wallEntryClass = WallEntry::class;
 
     /**
      * Flag for Entry Form to set this content to public
@@ -57,12 +55,6 @@ class CalendarEntry extends ContentActiveRecord implements Searchable
      * @inheritdoc
      */
     public $managePermission = ManageEntry::class;
-
-    /**
-     * Times
-     */
-    public $start_time;
-    public $end_time;
 
     /**
      * Participation Modes
@@ -84,9 +76,7 @@ class CalendarEntry extends ContentActiveRecord implements Searchable
     {
         parent::init();
 
-        /**
-         * Default participiation Mode
-         */
+        // Default participiation Mode
         $this->participation_mode = 2;
     }
 
@@ -111,10 +101,9 @@ class CalendarEntry extends ContentActiveRecord implements Searchable
         return [
             [['title', 'start_datetime', 'end_datetime'], 'required'],
             ['color', 'string'],
-            [['start_time', 'end_time'], 'date', 'format' => 'php:H:i'],
-            [['start_datetime'], DbDateValidator::className(), 'format' => Yii::$app->params['formatter']['defaultDateFormat'], 'timeAttribute' => 'start_time'],
-            [['end_datetime'], DbDateValidator::className(), 'format' => Yii::$app->params['formatter']['defaultDateFormat'], 'timeAttribute' => 'end_time'],
-            [['is_public', 'all_day'], 'integer'],
+            [['start_datetime'], DbDateValidator::className()],
+            [['end_datetime'], DbDateValidator::className()],
+            [['all_day'], 'integer'],
             [['title'], 'string', 'max' => 200],
             [['participation_mode'], 'in', 'range' => [self::PARTICIPATION_MODE_ALL, self::PARTICIPATION_MODE_INVITE, self::PARTICIPATION_MODE_NONE]],
             [['end_datetime'], 'validateEndTime'],
@@ -131,12 +120,7 @@ class CalendarEntry extends ContentActiveRecord implements Searchable
             'id' => Yii::t('CalendarModule.base', 'ID'),
             'title' => Yii::t('CalendarModule.base', 'Title'),
             'description' => Yii::t('CalendarModule.base', 'Description'),
-            'start_datetime' => Yii::t('CalendarModule.base', 'Start Date'),
-            'end_datetime' => Yii::t('CalendarModule.base', 'End Date'),
-            'start_time' => Yii::t('CalendarModule.base', 'Start Time'),
-            'end_time' => Yii::t('CalendarModule.base', 'End Time'),
             'all_day' => Yii::t('CalendarModule.base', 'All Day'),
-            'is_public' => Yii::t('CalendarModule.base', 'Public'),
             'participation_mode' => Yii::t('CalendarModule.base', 'Participation Mode'),
         ];
     }
@@ -150,40 +134,18 @@ class CalendarEntry extends ContentActiveRecord implements Searchable
      */
     public function validateEndTime($attribute, $params)
     {
-        if (new \DateTime($this->start_datetime) >= new \DateTime($this->end_datetime)) {
+        if (new DateTime($this->start_datetime) >= new DateTime($this->end_datetime)) {
             $this->addError($attribute, Yii::t('CalendarModule.base', "End time must be after start time!"));
         }
     }
 
-    public function afterFind()
-    {
-        parent::afterFind();
-
-        // Load form only attributes
-        $this->is_public = $this->content->visibility;
-        $this->start_time = Yii::$app->formatter->asTime($this->start_datetime, 'php:H:i');
-        $this->end_time = Yii::$app->formatter->asTime($this->end_datetime, 'php:H:i');
-    }
-
-    public function beforeDelete()
-    {
-
-        foreach (CalendarEntryParticipant::findAll(['calendar_entry_id' => $this->id]) as $participant) {
-            $participant->delete();
-        }
-
-        return parent::beforeDelete();
-    }
-
     public function beforeSave($insert)
     {
-        $this->content->visibility = $this->is_public ? Content::VISIBILITY_PUBLIC : Content::VISIBILITY_PRIVATE;
-
-        $startDateTime = new \DateTime($this->start_datetime);
-        $endDateTime = new \DateTime($this->end_datetime);
+        $startDateTime = new DateTime($this->start_datetime);
+        $endDateTime = new DateTime($this->end_datetime);
 
         // Check is a full day span
-        if ($this->all_day == 0 && \humhub\modules\calendar\Utils::isFullDaySpan($startDateTime, $endDateTime)) {
+        if ($this->all_day == 0 && CalendarUtils::isFullDaySpan($startDateTime, $endDateTime)) {
             $this->all_day = 1;
         }
 
@@ -195,9 +157,24 @@ class CalendarEntry extends ContentActiveRecord implements Searchable
         return parent::beforeSave($insert);
     }
 
-    public function inviteParticipant($user)
+    public function afterSave($insert, $changedAttributes)
     {
-        $this->setParticipant($user, CalendarEntryParticipant::PARTICIPATION_STATE_INVITED);
+        parent::afterSave($insert, $changedAttributes);
+
+        if ($insert) {
+            $this->setParticipant(Yii::$app->user->getIdentity());
+        }
+
+        return;
+    }
+
+    public function beforeDelete()
+    {
+        foreach (CalendarEntryParticipant::findAll(['calendar_entry_id' => $this->id]) as $participant) {
+            $participant->delete();
+        }
+
+        return parent::beforeDelete();
     }
 
     public function setParticipant($user, $state = CalendarEntryParticipant::PARTICIPATION_STATE_ACCEPTED)
@@ -229,17 +206,6 @@ class CalendarEntry extends ContentActiveRecord implements Searchable
         return CalendarEntryParticipant::findOne(['user_id' => $user->id, 'calendar_entry_id' => $this->id]);
     }
 
-    public function afterSave($insert, $changedAttributes)
-    {
-        parent::afterSave($insert, $changedAttributes);
-
-        if ($insert) {
-            $this->setParticipant(Yii::$app->user->getIdentity());
-        }
-
-        return;
-    }
-
     /**
      * Returns array for FullCalendar Javascript
      *
@@ -257,13 +223,17 @@ class CalendarEntry extends ContentActiveRecord implements Searchable
             $end = $endDateTime->format('Y-m-d');
         }
 
+        if(!Yii::$app->user->isGuest) {
+            Yii::$app->formatter->timeZone = Yii::$app->user->getIdentity()->time_zone;
+        }
+
         return [
             'id' => $this->id,
             'title' => $this->title,
             'editable' => $this->content->canWrite(),
             'backgroundColor' => $this->color,
             'allDay' => $this->all_day,
-            'updateUrl' => $this->content->container->createUrl('/calendar/entry/edit-ajax', ['id' => $this->id, 'end_datetime' => '-end-', 'start_datetime' => '-start-', 'fullCalendar' => '1']),
+            'updateUrl' => $this->content->container->createUrl('/calendar/entry/edit-ajax', ['id' => $this->id]),
             'viewUrl' => $this->content->container->createUrl('/calendar/entry/view', ['id' => $this->id, 'cal' => '1']),
             'start' => Yii::$app->formatter->asDatetime($this->start_datetime, 'php:c'),
             'end' => $end,
@@ -411,17 +381,6 @@ class CalendarEntry extends ContentActiveRecord implements Searchable
             'title' => $this->title,
             'description' => $this->description,
         ];
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function cutTime()
-    {
-        $this->start_datetime = preg_replace('/\d{2}:\d{2}:\d{2}$/', '', $this->start_datetime);
-        $this->end_datetime = preg_replace('/\d{2}:\d{2}:\d{2}$/', '', $this->end_datetime);
-        $this->start_time = '00:00';
-        $this->end_time = '23:59';
     }
 
     /**